+++ openfisca_core/parameters/parameter_scale_bracket.py
@@ -1,0 +1,9 @@
+from openfisca_core import parameters
+
+
+class ParameterScaleBracket(parameters.ParameterNode):
+    """
+    A parameter scale bracket.
+    """
+
+    _allowed_keys = set(['amount', 'threshold', 'rate', 'average_rate', 'base'])
