+++ openfisca_core/parameters/values_history.py
@@ -1,0 +1,9 @@
+from openfisca_core import parameters
+
+
+class ValuesHistory(parameters.Parameter):
+    """
+    Only for backward compatibility.
+    """
+
+    pass
